agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lognog-in"
version = "0.1.0"
description = "LogNog In - Lightweight log shipping agent for LogNog"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Machine King Labs"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
]
keywords = ["logging", "monitoring", "siem", "agent", "file-integrity"]

dependencies = [
    "watchdog>=3.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "pystray>=0.19.0",
    "pillow>=10.0.0",
    "appdirs>=1.4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pyinstaller>=6.0.0",
]

[project.scripts]
lognog-in = "lognog_in.main:main"

[project.urls]
Homepage = "https://github.com/machinekinglabs/lognog"
Documentation = "https://github.com/machinekinglabs/lognog/docs"
Repository = "https://github.com/machinekinglabs/lognog"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
